isolando componentes
