--- components/common/Sections.tsx
@@ -1,87 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-import { Link as LinkButton } from "./Link";
-import Link from "next/link";
-import sections, { storeSections } from "../../utils/sections";
-import { useRouter } from "next/router";
-
-interface Props {
-  mobile?: boolean;
-}
-
-interface ContainerProps {
-  mobile?: boolean;
-  width: string;
-}
-
-const Navs = styled.div<ContainerProps>`
-  display: flex;
-   @media only screen and (max-width: 600px) {
-      display: block;
-        width: ${(props) => props.width};
-
-  }
-`;
-const Container = styled.div<ContainerProps>`
-  display: flex;
-  flex-direction: ${(props) => (props.mobile ? "column" : "row")};
-  justify-content: ${(props) => (props.mobile ? "space-around" : "")};
-  height: ${(props) => (props.mobile ? "100%" : "")};
-  align-items: center !important;
-  width: ${(props) => props.width};
-
-  a {
-    margin-left: inherit;
-      align-items: flex-start !important;
-    margin-right: ${(props) => (props.mobile ? "" : "32px")};
-    z-index: 1;
-  }
-  a:after {
-    display: block;
-    content: "";
-    border-bottom: solid 3px #be42d7;
-    transform: scaleX(0);
-    transition: transform 0.2s ease-in-out;
-  }
-
-  a:hover:after {
-    transform: scaleX(1);
-  }
-
-  @media only screen and (max-width: 600px) {
-    ${(props) => (props.mobile ? "" : "display: none")};
-  }
-
-  .right {
-    padding-left: 40px;
-    margin-right: 0px;
-  }
-`;
-
-const Sections: React.FC<Props> = ({ mobile }: Props) => {
-  const router = useRouter();
-  const isStorePath = router.pathname.includes("/marketplace") ;
-  const routes = isStorePath ? sections : sections;
-
-  return (
-    <Container width={isStorePath ? "" : "100%"} mobile={mobile}>
-      <Navs width={isStorePath ? "" : "100%"} >
-      {routes.map((section) => (
-        <Link key={section.to} href={section.to}>
-          <a target={section.title === "White Paper" ? "_blank" : ""}>
-            {section.title}
-          </a>
-        </Link>
-      ))}
-      </Navs>
-      {isStorePath ? null : (
-        <LinkButton className="right" href="/marketplace">
-          MARKETPLACE
-        </LinkButton>
-      )}
-    </Container>
-  );
-};
-
-export default Sections;
