+++ pages/careers/index.tsx
@@ -3,7 +3,7 @@ import Career from '../../components/career/Career';
 import Header from '../../components/common/Header';
 import Ruler from '../../components/common/Ruler';
 import Footer from '../../components/common/Footer';
-import Sections from '../../components/common/Sections';
+import Sections from '../../components/common/Sections/Sections';
 import { Link } from '../../components/common/Link';
 import Logo from '../../components/common/Logo';
 import { Container } from '../../styles/Home';
